requirements.txt
yfinance
google-genai
streamlit
pydantic
python-dotenv
